Sketch out 'shell' approach
