+++ src/screens/PostThread/components/ThreadItemTreePost.tsx
@@ -44,6 +44,7 @@ import {RichText} from '#/components/RichText'
 import * as Skele from '#/components/Skeleton'
 import {SubtleHover} from '#/components/SubtleHover'
 import {Text} from '#/components/Typography'
+import {GalleryBleed} from '#/components/images/Gallery'
 
 /**
  * Mimic the space in PostMeta
@@ -129,33 +130,35 @@ const ThreadItemTreePostOuterWrapper = memo(
     const indents = Math.max(0, item.ui.indent - 1)
 
     return (
-      <View
-        style={[
-          a.flex_row,
-          item.ui.indent === 1 &&
-            !item.ui.showParentReplyLine && [
-              a.border_t,
-              t.atoms.border_contrast_low,
-            ],
-        ]}>
-        {Array.from(Array(indents)).map((_, n: number) => {
-          const isSkipped = item.ui.skippedIndentIndices.has(n)
-          return (
-            <View
-              key={`${item.value.post.uri}-padding-${n}`}
-              style={[
+      <GalleryBleed>
+        <View
+          style={[
+            a.flex_row,
+            item.ui.indent === 1 &&
+              !item.ui.showParentReplyLine && [
+                a.border_t,
                 t.atoms.border_contrast_low,
-                {
-                  borderRightWidth: isSkipped ? 0 : REPLY_LINE_WIDTH,
-                  width: TREE_INDENT + TREE_AVI_WIDTH / 2,
-                  left: 1,
-                },
-              ]}
-            />
-          )
-        })}
-        {children}
-      </View>
+              ],
+          ]}>
+          {Array.from(Array(indents)).map((_, n: number) => {
+            const isSkipped = item.ui.skippedIndentIndices.has(n)
+            return (
+              <View
+                key={`${item.value.post.uri}-padding-${n}`}
+                style={[
+                  t.atoms.border_contrast_low,
+                  {
+                    borderRightWidth: isSkipped ? 0 : REPLY_LINE_WIDTH,
+                    width: TREE_INDENT + TREE_AVI_WIDTH / 2,
+                    left: 1,
+                  },
+                ]}
+              />
+            )
+          })}
+          {children}
+        </View>
+      </GalleryBleed>
     )
   },
 )
